requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-braces==1.14.0
django-jazzmin==2.2.6
django-model-changes-py3==0.15.1
django-nested-inline==0.4.2
django-redis==4.12.1
django-simple-history==2.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
importlib-metadata==1.7.0
jdcal==1.4.1
kombu==4.6.11
openpyxl==3.0.4
packaging==20.4
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
whitenoise==5.2.0
zipp==3.1.0
